+++ Test/Acceptance_Tests/Acceptance_Test_Linux_App/test.sh
@@ -1,0 +1,12 @@
+#! /bin/bash
+
+pushd .
+
+# Switch current/working directory to here
+cd "${0%/*}"
+
+cd Build
+
+ctest .. --verbose
+
+popd